requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.1
django-jazzmin==3.0.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.6.4
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
twilio==9.8.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
yarl==1.20.1
